+++ exchanges_wrapper/exch_srv.py
@@ -16,13 +16,13 @@ import grpc
 from google.protobuf import json_format
 #
 from exchanges_wrapper import events, errors, api_pb2, api_pb2_grpc
-from exchanges_wrapper.client import Client
+from exchanges_wrapper.client import Client, STATUS_TIMEOUT
 from exchanges_wrapper.definitions import Side, OrderType, TimeInForce, ResponseType
-from exchanges_wrapper.c_structures import OrderUpdateEvent, OrderTradesEvent
+from exchanges_wrapper.c_structures import OrderUpdateEvent, OrderTradesEvent, REST_RATE_LIMIT_INTERVAL
 from exchanges_wrapper import WORK_PATH, CONFIG_FILE, LOG_FILE
 #
 HEARTBEAT = 1  # Sec
-MAX_QUEUE_SIZE = 50
+MAX_QUEUE_SIZE = 100
 #
 logger = logging.getLogger(__name__)
 formatter = logging.Formatter(fmt="[%(asctime)s: %(levelname)s] %(message)s")
@@ -106,6 +106,7 @@ class OpenClient:
             self.name = _account_name
             self.real_market = not account[2]
             self.client = Client(*account)
+            self.ts_rlc = time.time()
             OpenClient.open_clients.append(self)
         else:
             raise UserWarning(f"Account {_account_name} not registered into {WORK_PATH}/config/exch_srv_cfg.toml")
@@ -134,6 +135,14 @@ class OpenClient:
 class Martin(api_pb2_grpc.MartinServicer):
     rate_limit_reached_time = None
     rate_limiter = None
+
+    async def rate_limit_control(self, client, _call_from='default'):
+        if client.client.exchange == 'bitfinex':
+            rate_limit_interval = REST_RATE_LIMIT_INTERVAL.get(client.client.exchange, {}).get(_call_from, 0)
+            ts_diff = time.time() - client.ts_rlc
+            if ts_diff < rate_limit_interval:
+                sleep_duration = rate_limit_interval - ts_diff
+                await asyncio.sleep(sleep_duration)
 
     async def OpenClientConnection(self, request: api_pb2.OpenClientConnectionRequest,
                                    _context: grpc.aio.ServicerContext) -> api_pb2.OpenClientConnectionId:
@@ -196,6 +205,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                               _context: grpc.aio.ServicerContext) -> api_pb2.FetchServerTimeResponse:
         open_client = OpenClient.get_client(request.client_id)
         client = open_client.client
+        await self.rate_limit_control(open_client)
         try:
             res = await client.fetch_server_time()
         except Exception as ex:
@@ -203,6 +213,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             _context.set_details(f"{ex}")
             _context.set_code(grpc.StatusCode.UNKNOWN)
         else:
+            open_client.ts_rlc = time.time()
             server_time = res.get('serverTime')
             return api_pb2.FetchServerTimeResponse(server_time=server_time)
 
@@ -225,10 +236,9 @@ class Martin(api_pb2_grpc.MartinServicer):
                               _context: grpc.aio.ServicerContext) -> api_pb2.FetchOpenOrdersResponse:
         open_client = OpenClient.get_client(request.client_id)
         client = open_client.client
-        # message list
         response = api_pb2.FetchOpenOrdersResponse()
-        # Nested dict
         response_order = api_pb2.FetchOpenOrdersResponse.Order()
+        await self.rate_limit_control(open_client)
         try:
             res = await client.fetch_open_orders(request.trade_id, request.symbol)
         except asyncio.CancelledError:
@@ -249,23 +259,25 @@ class Martin(api_pb2_grpc.MartinServicer):
             _context.set_code(grpc.StatusCode.UNKNOWN)
         else:
             # logger.info(f"FetchOpenOrders.res: {res}")
+            open_client.ts_rlc = time.time()
             active_orders = []
             for order in res:
-                active_orders.append(order['orderId'])
+                order_id = order['orderId']
+                active_orders.append(order_id)
                 new_order = json_format.ParseDict(order, response_order)
                 # logger.debug(f"FetchOpenOrders.new_order: {new_order}")
                 response.items.append(new_order)
                 if client.exchange == 'bitfinex':
-                    client.active_orders.update(
-                        {order['orderId']:
-                            {'filledTime': int(),
-                             'origQty': order['origQty'],
-                             'executedQty': order['executedQty'],
-                             'lastEvent': (),
-                             'cancelled': False
-                             }
-                         }
-                    )
+                    if order_id in client.active_orders:
+                        client.active_orders[order_id]['executedQty'] = order['executedQty']
+                    else:
+                        client.active_orders[order_id] = {
+                            'lifeTime': int(time.time()) + 60 * STATUS_TIMEOUT,
+                            'origQty': order['origQty'],
+                            'executedQty': order['executedQty'],
+                            'lastEvent': (),
+                            'cancelled': False
+                        }
             if client.exchange == 'bitfinex':
                 client.active_orders_clear(active_orders)
         response.rate_limiter = Martin.rate_limiter
@@ -277,6 +289,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         _queue = client.on_order_update_queues.get(request.trade_id)
         response = api_pb2.FetchOrderResponse()
+        await self.rate_limit_control(open_client)
         try:
             res = await client.fetch_order(
                 request.trade_id,
@@ -290,6 +303,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         except Exception as _ex:
             logger.error(f"FetchOrders for {open_client.name}: {request.symbol} exception: {_ex}")
         else:
+            open_client.ts_rlc = time.time()
             if _queue and request.filled_update_call:
                 if res.get('status') == 'FILLED':
                     event = OrderUpdateEvent(res)
@@ -343,58 +357,61 @@ class Martin(api_pb2_grpc.MartinServicer):
         return response
 
     async def FetchExchangeInfoSymbol(self, request: api_pb2.MarketRequest,
-                                      _context: grpc.aio.ServicerContext
-                                      ) -> api_pb2.FetchExchangeInfoSymbolResponse:
-        client = OpenClient.get_client(request.client_id).client
+                                      _context: grpc.aio.ServicerContext) -> api_pb2.FetchExchangeInfoSymbolResponse:
+        open_client = OpenClient.get_client(request.client_id)
+        client = open_client.client
         response = api_pb2.FetchExchangeInfoSymbolResponse()
         exchange_info = await client.fetch_exchange_info()
         exchange_info_symbol = {}
+        await self.rate_limit_control(open_client)
         try:
             exchange_info_symbol = next(item for item in exchange_info.get('symbols')
                                         if item["symbol"] == request.symbol)
         except StopIteration:
             logger.info("FetchExchangeInfoSymbol.exchange_info_symbol: None")
         # logger.info(f"exchange_info_symbol: {exchange_info_symbol}")
+        open_client.ts_rlc = time.time()
         filters_res = exchange_info_symbol.pop('filters', [])
         json_format.ParseDict(exchange_info_symbol, response)
         # logger.info(f"filters: {filters_res}")
         filters = response.filters
         for _filter in filters_res:
-            if _filter.get('filterType') == 'PRICE_FILTER':
+            filter_type = _filter.get('filterType')
+            if filter_type == 'PRICE_FILTER':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.PriceFilter()
                 filters.price_filter.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif 'PERCENT_PRICE' in _filter.get('filterType'):
-                if _filter.get('filterType') == 'PERCENT_PRICE_BY_SIDE':
+            elif 'PERCENT_PRICE' in filter_type:
+                if filter_type == 'PERCENT_PRICE_BY_SIDE':
                     _filter['multiplierUp'] = _filter['bidMultiplierUp']
                     _filter['multiplierDown'] = _filter['bidMultiplierDown']
-                    _filter.pop('bidMultiplierUp')
-                    _filter.pop('bidMultiplierDown')
-                    _filter.pop('askMultiplierUp')
-                    _filter.pop('askMultiplierDown')
+                    del _filter['bidMultiplierUp']
+                    del _filter['bidMultiplierDown']
+                    del _filter['askMultiplierUp']
+                    del _filter['askMultiplierDown']
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.PercentPrice()
                 filters.percent_price.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'LOT_SIZE':
+            elif filter_type == 'LOT_SIZE':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.LotSize()
                 filters.lot_size.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'MIN_NOTIONAL':
+            elif filter_type == 'MIN_NOTIONAL':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.MinNotional()
                 filters.min_notional.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'NOTIONAL':
+            elif filter_type == 'NOTIONAL':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.Notional()
                 filters.notional.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'ICEBERG_PARTS':
+            elif filter_type == 'ICEBERG_PARTS':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.IcebergParts()
                 filters.iceberg_parts.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'MARKET_LOT_SIZE':
+            elif filter_type == 'MARKET_LOT_SIZE':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.MarketLotSize()
                 filters.market_lot_size.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'MAX_NUM_ORDERS':
+            elif filter_type == 'MAX_NUM_ORDERS':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.MaxNumOrders()
                 filters.max_num_orders.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'MAX_NUM_ICEBERG_ORDERS':
+            elif filter_type == 'MAX_NUM_ICEBERG_ORDERS':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.MaxNumIcebergOrders()
                 filters.max_num_iceberg_orders.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
-            elif _filter.get('filterType') == 'MAX_POSITION':
+            elif filter_type == 'MAX_POSITION':
                 new_filter_template = api_pb2.FetchExchangeInfoSymbolResponse.Filters.MaxPosition()
                 filters.max_position.CopyFrom(json_format.ParseDict(_filter, new_filter_template))
         return response
@@ -406,7 +423,9 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         response = api_pb2.FetchAccountBalanceResponse()
         response_balance = api_pb2.FetchAccountBalanceResponse.Balances()
+        await self.rate_limit_control(open_client)
         account_information = await client.fetch_account_information(request.trade_id, receive_window=None)
+        open_client.ts_rlc = time.time()
         # Send only balances
         res = account_information.get('balances', [])
         # Create consolidated list of asset balances from SPOT and Funding wallets
@@ -428,6 +447,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         response = api_pb2.FetchFundingWalletResponse()
         response_balance = api_pb2.FetchFundingWalletResponse.Balances()
+        await self.rate_limit_control(open_client)
         res = []
         if client.exchange in ('bitfinex', 'okx') or (open_client.real_market and client.exchange == 'binance'):
             try:
@@ -436,6 +456,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                                                         receive_window=request.receive_window)
             except AttributeError:
                 logger.error("Can't get Funding Wallet balances")
+        open_client.ts_rlc = time.time()
         logger.debug(f"funding_wallet: {res}")
         for balance in res:
             new_balance = json_format.ParseDict(balance, response_balance)
@@ -444,10 +465,13 @@ class Martin(api_pb2_grpc.MartinServicer):
 
     async def FetchOrderBook(self, request: api_pb2.MarketRequest,
                              _context: grpc.aio.ServicerContext) -> api_pb2.FetchOrderBookResponse:
-        client = OpenClient.get_client(request.client_id).client
+        open_client = OpenClient.get_client(request.client_id)
+        client = open_client.client
         response = api_pb2.FetchOrderBookResponse()
+        await self.rate_limit_control(open_client)
         limit = 1 if client.exchange in ('bitfinex', 'okx') else 5
         res = await client.fetch_order_book(symbol=request.symbol, limit=limit)
+        open_client.ts_rlc = time.time()
         res_bids = res.get('bids', [])
         res_asks = res.get('asks', [])
         response.lastUpdateId = res.get('lastUpdateId')
@@ -460,25 +484,33 @@ class Martin(api_pb2_grpc.MartinServicer):
     async def FetchSymbolPriceTicker(
             self, request: api_pb2.MarketRequest,
             _context: grpc.aio.ServicerContext) -> api_pb2.FetchSymbolPriceTickerResponse:
-        client = OpenClient.get_client(request.client_id).client
+        open_client = OpenClient.get_client(request.client_id)
+        client = open_client.client
         response = api_pb2.FetchSymbolPriceTickerResponse()
+        await self.rate_limit_control(open_client)
         res = await client.fetch_symbol_price_ticker(symbol=request.symbol)
+        open_client.ts_rlc = time.time()
         json_format.ParseDict(res, response)
         return response
 
     async def FetchTickerPriceChangeStatistics(
             self, request: api_pb2.MarketRequest,
             _context: grpc.aio.ServicerContext) -> api_pb2.FetchTickerPriceChangeStatisticsResponse:
-        client = OpenClient.get_client(request.client_id).client
+        open_client = OpenClient.get_client(request.client_id)
+        client = open_client.client
         response = api_pb2.FetchTickerPriceChangeStatisticsResponse()
+        await self.rate_limit_control(open_client)
         res = await client.fetch_ticker_price_change_statistics(symbol=request.symbol)
+        open_client.ts_rlc = time.time()
         json_format.ParseDict(res, response)
         return response
 
     async def FetchKlines(self, request: api_pb2.FetchKlinesRequest,
                           _context: grpc.aio.ServicerContext) -> api_pb2.FetchKlinesResponse:
-        client = OpenClient.get_client(request.client_id).client
+        open_client = OpenClient.get_client(request.client_id)
+        client = open_client.client
         response = api_pb2.FetchKlinesResponse()
+        await self.rate_limit_control(open_client)
         try:
             res = await client.fetch_klines(symbol=request.symbol, interval=request.interval,
                                             start_time=None, end_time=None, limit=request.limit)
@@ -488,6 +520,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             logger.error(f"FetchKlines for {request.symbol} interval: {request.interval}, exception: {_ex}")
         else:
             # logger.info(f"FetchKlines.res: {res}")
+            open_client.ts_rlc = time.time()
             for candle in res:
                 response.klines.append(json.dumps(candle))
         return response
@@ -518,6 +551,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                 event_handler, _queue, client, request.trade_id, _event_type),
                 _event_type, exchange, request.trade_id)
         while True:
+            response.Clear()
             _event = await _queue.get()
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
@@ -542,6 +576,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                           ]
                 response.candle = json.dumps(candle)
                 yield response
+                _queue.task_done()
 
     async def FetchAccountTradeList(self, request: api_pb2.AccountTradeListRequest,
                                     _context: grpc.aio.ServicerContext) -> api_pb2.AccountTradeListResponse:
@@ -549,6 +584,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         response = api_pb2.AccountTradeListResponse()
         response_trade = api_pb2.AccountTradeListResponse.Trade()
+        await self.rate_limit_control(open_client)
         try:
             res = await client.fetch_account_trade_list(
                 request.trade_id,
@@ -564,6 +600,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             logger.error(f"FetchAccountTradeList for {open_client.name}: {request.symbol} exception: {_ex}")
         else:
             # logger.info(f"FetchAccountTradeList: {res}")
+            open_client.ts_rlc = time.time()
             for trade in res:
                 trade_order = json_format.ParseDict(trade, response_trade)
                 response.items.append(trade_order)
@@ -586,6 +623,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         client.events.register_event(functools.partial(event_handler, _queue, client, request.trade_id, _event_type),
                                      _event_type, client.exchange, request.trade_id)
         while True:
+            response.Clear()
             _event = await _queue.get()
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
@@ -599,13 +637,14 @@ class Martin(api_pb2_grpc.MartinServicer):
                               'event_time': _event.event_time}
                 json_format.ParseDict(ticker_24h, response)
                 yield response
+                _queue.task_done()
 
     async def OnOrderBookUpdate(self, request: api_pb2.MarketRequest,
                                 _context: grpc.aio.ServicerContext) -> api_pb2.FetchOrderBookResponse:
         response = api_pb2.FetchOrderBookResponse()
         open_client = OpenClient.get_client(request.client_id)
         client = open_client.client
-        _queue = asyncio.Queue(MAX_QUEUE_SIZE * 2)
+        _queue = asyncio.LifoQueue(MAX_QUEUE_SIZE * 5)
         client.stream_queue[request.trade_id] |= {_queue}
         if client.exchange == 'okx':
             _symbol = client.symbol_to_okx(request.symbol)
@@ -617,19 +656,21 @@ class Martin(api_pb2_grpc.MartinServicer):
         client.events.register_event(functools.partial(event_handler, _queue, client, request.trade_id, _event_type),
                                      _event_type, client.exchange, request.trade_id)
         while True:
+            response.Clear()
             _event = await _queue.get()
+            while not _queue.empty():
+                _queue.get_nowait()
+                _queue.task_done()
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
                 logger.info(f"OnOrderBookUpdate: Stop loop for {open_client.name}: {request.symbol}")
                 return
             else:
-                response.Clear()
                 response.lastUpdateId = _event.last_update_id
-                for bid in _event.bids:
-                    response.bids.append(json.dumps(bid))
-                for ask in _event.asks:
-                    response.asks.append(json.dumps(ask))
+                [response.bids.append(json.dumps(bid)) for bid in _event.bids]
+                [response.asks.append(json.dumps(ask)) for ask in _event.asks]
                 yield response
+                _queue.task_done()
 
     async def OnFundsUpdate(self, request: api_pb2.OnFundsUpdateRequest,
                             _context: grpc.aio.ServicerContext) -> api_pb2.OnFundsUpdateResponse:
@@ -642,6 +683,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             event_handler, _queue, client, request.trade_id, 'outboundAccountPosition'),
             'outboundAccountPosition')
         while True:
+            response.Clear()
             _event = await _queue.get()
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
@@ -651,6 +693,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                 # logger.debug(f"OnFundsUpdate: {client.exchange}:{_event.balances.items()}")
                 response.funds = json.dumps(_event.balances)
                 yield response
+                _queue.task_done()
 
     async def OnBalanceUpdate(self, request: api_pb2.MarketRequest,
                               _context: grpc.aio.ServicerContext) -> api_pb2.OnBalanceUpdateResponse:
@@ -663,20 +706,25 @@ class Martin(api_pb2_grpc.MartinServicer):
             client.events.register_user_event(functools.partial(
                 event_handler, _queue, client, request.trade_id, 'balanceUpdate'), 'balanceUpdate')
         while True:
+            response.Clear()
             try:
-                _event = await asyncio.wait_for(_queue.get(), timeout=HEARTBEAT * 10)
+                _event = await asyncio.wait_for(_queue.get(), timeout=HEARTBEAT * 30)
+                _get_event_from_queue = True
             except asyncio.TimeoutError:
                 _event = None
+                _get_event_from_queue = False
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
                 logger.info(f"OnBalanceUpdate: Stop user stream for {open_client.name}:{request.symbol}")
                 return
             if client.exchange in ('bitfinex', 'huobi'):
+                await self.rate_limit_control(open_client)
                 try:
                     balance = await client.fetch_ledgers(request.symbol)
                 except Exception as _ex:
                     logger.warning(f"OnBalanceUpdate: for {open_client.name}:{request.symbol}: {_ex}")
                 else:
+                    open_client.ts_rlc = time.time()
                     if balance:
                         _event = client.events.wrap_event(balance)
             if isinstance(_event, events.BalanceUpdateWrapper):
@@ -691,6 +739,8 @@ class Martin(api_pb2_grpc.MartinServicer):
                     }
                     response.balance = json.dumps(balance)
                     yield response
+                    if _get_event_from_queue:
+                        _queue.task_done()
 
     async def OnOrderUpdate(self, request: api_pb2.MarketRequest,
                             _context: grpc.aio.ServicerContext) -> api_pb2.SimpleResponse:
@@ -704,6 +754,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             event_handler, _queue, client, request.trade_id, 'executionReport'),
             'executionReport')
         while True:
+            response.Clear()
             _event = await _queue.get()
             if isinstance(_event, str) and _event == request.trade_id:
                 client.stream_queue.get(request.trade_id, set()).discard(_queue)
@@ -716,6 +767,7 @@ class Martin(api_pb2_grpc.MartinServicer):
                 response.success = True
                 response.result = json.dumps(str(event))
                 yield response
+                _queue.task_done()
 
     async def CreateLimitOrder(self, request: api_pb2.CreateLimitOrderRequest,
                                _context: grpc.aio.ServicerContext) -> api_pb2.CreateLimitOrderResponse:
@@ -794,6 +846,7 @@ class Martin(api_pb2_grpc.MartinServicer):
         response.success = False
         open_client = OpenClient.get_client(request.client_id)
         client = open_client.client
+        await self.rate_limit_control(open_client)
         try:
             res = await client.transfer_to_master(symbol=request.symbol, quantity=request.amount)
         except errors.HTTPError as ex:
@@ -806,6 +859,7 @@ class Martin(api_pb2_grpc.MartinServicer):
             _context.set_details(f"{ex}")
             _context.set_code(grpc.StatusCode.UNKNOWN)
         else:
+            open_client.ts_rlc = time.time()
             if res and res.get("txnId"):
                 response.success = True
             response.result = json.dumps(res)
@@ -826,8 +880,8 @@ class Martin(api_pb2_grpc.MartinServicer):
             _market_stream_count = sum(len(k) for k in ([list(i.get(request.trade_id, []))
                                                          for i in list(client.events.registered_streams.values())]))
         logger.info(f"Start WS streams for {open_client.name}")
-        asyncio.create_task(client.start_market_events_listener(request.trade_id))
-        asyncio.create_task(client.start_user_events_listener(request.trade_id, request.symbol))
+        await client.start_market_events_listener(request.trade_id)
+        await client.start_user_events_listener(request.trade_id, request.symbol)
         response.success = True
         return response
 
@@ -860,15 +914,15 @@ async def stop_stream(client, trade_id):
     await client.stop_events_listener(trade_id)
     client.events.unregister(client.exchange, trade_id)
     [await _queue.put(trade_id) for _queue in client.stream_queue.get(trade_id, [])]
+    await asyncio.sleep(0)
     client.on_order_update_queues.pop(trade_id, None)
     client.stream_queue.pop(trade_id, None)
     gc.collect(generation=2)
 
 
 async def event_handler(_queue, client, trade_id, _event_type, event):
-    _event = weakref.ref(event)
     try:
-        _queue.put_nowait(_event())
+        _queue.put_nowait(weakref.ref(event)())
     except asyncio.QueueFull:
         logger.warning(f"For {_event_type} asyncio queue full and wold be closed")
         client.stream_queue.get(trade_id, set()).discard(_queue)